int to double error fix

+++ lib/screens/payment/complete_payment.dart
@@ -79,7 +79,8 @@ class CompleteActionPayment extends StatelessWidget {
                       ),
                       Expanded(
                         child: CustomColumn(
-                          rate: (int.parse(paidAmount) - int.parse(totalAmount))
+                          rate: (double.parse(paidAmount) -
+                                  double.parse(totalAmount))
                               .toString(),
                           title: 'Change',
                         ),
